test: move test code into a separate module
This enables easier manual testing with `cargo expand`, and will enable easier
snapshot testing later.
